Divisors

+++ Module6.5/Ali_Baba_and_Puzzles.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main()
 {
-    int a,b,c,n;
-    scanf("%d %d %d %d",&a,&b,&c,&n);
+    long long int a,b,c,n;
+    scanf("%lld %lld %lld %lld",&a,&b,&c,&n);
     if(a+b*c==n || a+b-c==n|| a-b+c==n || a-b*c==n || a*b-c==n || a*b+c==n)
     {
         printf("YES");
